+++ dflib-echarts/src/main/java/org/dflib/echarts/CandlestickSeriesOpts.java
@@ -1,0 +1,12 @@
+package org.dflib.echarts;
+
+/**
+ * @since 1.0.0-M24
+ */
+public class CandlestickSeriesOpts extends CartesianSeriesOpts<CandlestickSeriesOpts> {
+
+    @Override
+    public ChartType getType() {
+        return ChartType.candlestick;
+    }
+}