+++ dflib-echarts/src/main/java/org/dflib/echarts/CartesianSeriesOpts.java
@@ -5,25 +5,8 @@ package org.dflib.echarts;
  */
 public abstract class CartesianSeriesOpts<SO extends CartesianSeriesOpts<SO>> extends SeriesOpts<SO> {
 
-    protected Label label;
     protected Integer xAxisIndex;
     protected Integer yAxisIndex;
-
-    /**
-     * @since 1.0.0-M22
-     */
-    public SO label(LabelPosition position) {
-        this.label = Label.of(position);
-        return (SO) this;
-    }
-
-    /**
-     * @since 1.0.0-M22
-     */
-    public SO label(Label label) {
-        this.label = label;
-        return (SO) this;
-    }
 
     /**
      * Sets an index of X axis to use for this Series. There can be one or more X axes, so this method allows to